Dead code removal

+++ components/com_finder/tmpl/search/default_results.php
@@ -90,7 +90,7 @@ use Joomla\CMS\Uri\Uri;
             <?php $start = (int) $this->pagination->limitstart + 1; ?>
             <?php $total = (int) $this->pagination->total; ?>
             <?php $limit = (int) $this->pagination->limit * $this->pagination->pagesCurrent; ?>
-            <?php $limit = (int) min($limit, $total); ?>
+            <?php $limit = min($limit, $total); ?>
             <?php echo Text::sprintf('COM_FINDER_SEARCH_RESULTS_OF', $start, $limit, $total); ?>
         </div>
     <?php endif; ?>
